requirements.txt
cassandra-driver=2.5.1